pyproject.toml
[tool.poetry]
name = "detectionmetrics"
version = "2.2.0"
description = "Tools for evaluating segmentation and object detection models"
authors = ["JdeRobot", "d.pascualhe <d.pascualhe@gmail.com>"]
readme = "README.md"
license = "LICENSE"

[tool.poetry.dependencies]
python = "^3.10"
tqdm = "^4.65.0"
pandas = "^2.2.3"
PyYAML = "^6.0.2"
pyarrow = "^18.0.0"
pillow = "^11.0.0"
numpy = "1.26.4"
opencv-python-headless = "^4.10.0.84"
scikit-learn = "^1.6.0"
open3d = "^0.19.0"
addict = "^2.4.0"
matplotlib = "^3.6.0"
click = "^8.1.8"
tensorboard = "^2.18.0"
pycocotools = { version = "^2.0.7", markers = "sys_platform != 'win32'" }
pycocotools-windows = { version = "^2.0.0.2", markers = "sys_platform == 'win32'" }
Streamlit = "1.46.0"
streamlit-image-select = "^0.6.0"
supervision = "^0.18.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pylint = "^3.3.1"
ipykernel = "^6.29.5"

[tool.poetry.group.docs.dependencies]
sphinx = "^8.1.3"
sphinx-rtd-theme = "^3.0.2"

[tool.poetry.group.test.dependencies]
pytest = "^8.0.0"

[tool.poetry.scripts]
dm_evaluate = "detectionmetrics.cli.evaluate:evaluate"
dm_batch = "detectionmetrics.cli.batch:batch"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
